+++ src/main/php/net/stubbles/webapp/WebApp.php
@@ -10,6 +10,7 @@
 namespace net\stubbles\webapp;
 use net\stubbles\input\web\WebRequest;
 use net\stubbles\ioc\App;
+use net\stubbles\lang\errorhandler\ExceptionLogger;
 use net\stubbles\webapp\ioc\IoBindingModule;
 use net\stubbles\webapp\response\Response;
 use net\stubbles\webapp\response\ResponseNegotiator;
@@ -38,22 +39,31 @@ abstract class WebApp extends App
      * @type  Routing
      */
     private $routing;
+    /**
+     * logger for logging uncatched exceptions
+     *
+     * @type  ExceptionLogger
+     */
+    private $exceptionLogger;
 
     /**
      * constructor
      *
      * @param  WebRequest          $request             request data container
      * @param  ResponseNegotiator  $responseNegotiator  negoatiates based on request
-     * @param  Routing             $routing
+     * @param  Routing             $routing             routes to logic based on request
+     * @param  ExceptionLogger     $exceptionLogger     logs uncatched exceptions
      * @Inject
      */
     public function __construct(WebRequest $request,
                                 ResponseNegotiator $responseNegotiator,
-                                Routing $routing)
+                                Routing $routing,
+                                ExceptionLogger $exceptionLogger)
     {
         $this->request            = $request;
         $this->responseNegotiator = $responseNegotiator;
         $this->routing            = $routing;
+        $this->exceptionLogger    = $exceptionLogger;
     }
 
     /**
@@ -97,6 +107,7 @@ abstract class WebApp extends App
                 }
             }
         } catch (\Exception $e) {
+            $this->exceptionLogger->log($e);
             $response->internalServerError($e->getMessage());
         }
     }