+++ src/test/php/net/stubbles/webapp/WebAppTestCase.php
@@ -60,6 +60,12 @@ class WebAppTestCase extends \PHPUnit_Framework_TestCase
      * @type  Routing
      */
     private $routing;
+    /**
+     * mocked exception logger
+     *
+     * @type  \PHPUnit_Framework_MockObject_MockObject
+     */
+    private $mockExceptionLogger;
 
     /**
      * set up test environment
@@ -83,11 +89,15 @@ class WebAppTestCase extends \PHPUnit_Framework_TestCase
         $this->routing = $this->getMockBuilder('net\stubbles\webapp\Routing')
                               ->disableOriginalConstructor()
                               ->getMock();
+        $this->mockExceptionLogger = $this->getMockBuilder('net\stubbles\lang\errorhandler\ExceptionLogger')
+                                          ->disableOriginalConstructor()
+                                          ->getMock();
         $this->webApp  = $this->getMock('net\stubbles\webapp\TestWebApp',
                                         array('configureRouting'),
                                         array($this->mockRequest,
                                               $mockResponseNegotiator,
-                                              $this->routing
+                                              $this->routing,
+                                              $this->mockExceptionLogger
                                         )
                          );
     }
@@ -205,16 +215,20 @@ class WebAppTestCase extends \PHPUnit_Framework_TestCase
         $mockRoute->expects($this->once())
                   ->method('switchToHttps')
                   ->will($this->returnValue(false));
+        $exception = new \Exception('some error');
         $mockRoute->expects($this->once())
                   ->method('applyPreInterceptors')
                   ->with($this->equalTo($this->mockRequest),
                          $this->equalTo($this->mockResponse)
                     )
-                  ->will($this->throwException(new \Exception('some error')));
+                  ->will($this->throwException($exception));
         $mockRoute->expects($this->never())
                   ->method('process');
         $mockRoute->expects($this->never())
                   ->method('applyPostInterceptors');
+        $this->mockExceptionLogger->expects($this->once())
+                                  ->method('log')
+                                  ->with($this->equalTo($exception));
         $this->mockResponse->expects($this->once())
                            ->method('internalServerError')
                            ->with($this->equalTo('some error'));
@@ -266,14 +280,18 @@ class WebAppTestCase extends \PHPUnit_Framework_TestCase
                          $this->equalTo($this->mockResponse)
                     )
                   ->will($this->returnValue(true));
+        $exception = new \Exception('some error');
         $mockRoute->expects($this->once())
                   ->method('process')
                   ->with($this->equalTo($this->mockRequest),
                          $this->equalTo($this->mockResponse)
                     )
-                  ->will($this->throwException(new \Exception('some error')));
+                  ->will($this->throwException($exception));
         $mockRoute->expects($this->never())
                   ->method('applyPostInterceptors');
+        $this->mockExceptionLogger->expects($this->once())
+                                  ->method('log')
+                                  ->with($this->equalTo($exception));
         $this->mockResponse->expects($this->once())
                            ->method('internalServerError')
                            ->with($this->equalTo('some error'));
@@ -334,12 +352,16 @@ class WebAppTestCase extends \PHPUnit_Framework_TestCase
                          $this->equalTo($this->mockResponse)
                     )
                   ->will($this->returnValue(true));
+        $exception = new \Exception('some error');
         $mockRoute->expects($this->once())
                   ->method('applyPostInterceptors')
                   ->with($this->equalTo($this->mockRequest),
                          $this->equalTo($this->mockResponse)
                     )
-                  ->will($this->throwException(new \Exception('some error')));
+                  ->will($this->throwException($exception));
+        $this->mockExceptionLogger->expects($this->once())
+                                  ->method('log')
+                                  ->with($this->equalTo($exception));
         $this->mockResponse->expects($this->once())
                            ->method('internalServerError')
                            ->with($this->equalTo('some error'));
@@ -370,7 +392,8 @@ class WebAppTestCase extends \PHPUnit_Framework_TestCase
                                         array('configureRouting'),
                                         array($this->mockRequest,
                                               $mockResponseNegotiator,
-                                              $this->routing
+                                              $this->routing,
+                                              $this->mockExceptionLogger
                                         )
                          );
         $mockRoute = $this->createMockRoute();